src/webvis-flask/requirements.txt
flask
flask-cors
opencv-python
numpy
PyMySQL
python-dotenv
cryptography